+++ vsphere/assets/stubs/vim/event.pyi
@@ -1,0 +1,3 @@
+
+class Event:
+    pass
